requirements-backend.txt
fastapi
uvicorn
langchain
langgraph
openai
langchain_chroma
pypdf
pandas
tiktoken
azure-identity
python-dotenv
langchain-text-splitters
langchain-core
langchain_openai
langchain-community
PyJWT
cryptography
httpx
requests>=2.31.0
azure-search-documents
azure-storage-blob
azure-core